+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 1.0.1
+
+1. request method changed to POST
+2. test_mode configuration variable renamed to testMode
+3. added sslVerify configuration variable. 
+>This allows you to disable ssl validation (don't do this!) or use a custom SSL certificate from file system
+
 ## 0.1.3
 
 Optimize registerOrder input patam processing